team project/requirements.txt
Flask==2.1.1
flask-cors==3.0.10
python-dotenv==0.19.2
requests==2.27.1
openai==0.27.0
langchain==0.0.9
langchain-openai==0.0.9
langchain-community==0.0.9
PyMuPDF==1.19.6
spacy==3.2.0
